consolidated components into comonent folder
