+++ src/app/page.js
@@ -1,8 +1,8 @@
 import Image from "next/image";
 import styles from "./page.module.css";
-import Hero from "./hero/Hero";
-import Section1 from "./section1/page";
-import Section2 from "./section2/page";
+import Hero from "./components/hero/Hero";
+import Section1 from "./components/section1/Section1";
+import Section2 from "./components/section2/Section2";
 
 export default function Home() {
   return (
